+++ examples/libhello/src/test/ChildClassImpl.h
@@ -32,7 +32,7 @@ public:
     ~ChildClassImpl( ) = default;
     ::std::shared_ptr< ParentInterface > cast_to_parent( ) override;
     ::std::string get_name( ) override;
-    int32_t get_lucky_number( ) override;
+    int32_t get_lucky_number( ) const override;
     void set_lucky_number( const int32_t number ) override;
     void do_something_to_child_class(
         const ::std::shared_ptr< ::test::ChildClass >& input ) override;